+++ src/tx-params/convert-sell.js
@@ -3,6 +3,7 @@ import {TX_TYPE} from 'minterjs-tx';
 import SellTxData from '../tx-data/convert-sell.js';
 
 /**
+ * @deprecated
  * @constructor
  * @param {string} coinFrom
  * @param {string} coinTo
@@ -13,6 +14,9 @@ import SellTxData from '../tx-data/convert-sell.js';
  * @return {TxParams}
  */
 export default function SellTxParams({coinFrom, coinTo, sellAmount, minBuyAmount = 0, feeCoinSymbol, ...otherParams}) {
+    // eslint-disable-next-line no-console
+    console.warn('SellTxParams is deprecated');
+
     const txData = new SellTxData({
         coinToSell: coinFrom,
         coinToBuy: coinTo,