+++ src/tx-params/send.js
@@ -2,6 +2,7 @@ import {TX_TYPE} from 'minterjs-tx';
 import SendTxData from '../tx-data/send.js';
 
 /**
+ * @deprecated
  * @constructor
  * @param {string} address
  * @param {number|string} amount
@@ -11,6 +12,9 @@ import SendTxData from '../tx-data/send.js';
  * @return {TxParams}
  */
 export default function SendTxParams({address, amount = 0, coinSymbol, feeCoinSymbol, ...otherParams}) {
+    // eslint-disable-next-line no-console
+    console.warn('SendTxParams is deprecated');
+
     const txData = new SendTxData({
         to: address,
         coin: coinSymbol,