+++ src/tx.js
@@ -33,6 +33,12 @@ import {decodeTxData, ensureBufferData} from './tx-data/index.js';
  * @return {Tx}
  */
 export default function prepareSignedTx(txParams = {}, {privateKey} = {}) {
+    if (!privateKey && txParams.privateKey) {
+        privateKey = txParams.privateKey;
+        // eslint-disable-next-line no-console
+        console.warn('privateKey field in tx params is deprecated, pass it to the second parameter');
+    }
+    console.log({privateKey});
     const tx = prepareTx({...txParams, signatureType: 1}, {privateKey});
 
     tx.signatureData = makeSignature(tx, privateKey);